+++ plugins/tmpl-mxtag.js
@@ -16,14 +16,15 @@ let util = require('util');
 let chalk = require('chalk');
 let tmplParser = require('./tmpl-parser');
 let attrMap = require('./tmpl-attr-map');
+let sep = path.sep;
 let uncheckTags = { view: 1, include: 1, native: 1 };
-let cmdOutReg = /<%([!=@])([\s\S]*)%>/;
+let cmdOutReg = /^<%([!=@])([\s\S]*)%>$/;
 let tagReg = /\btag\s*=\s*"([^"]+)"/;
 let attrNameValueReg = /\s*([^=\/\s]+)\s*=\s*(["'])([\s\S]*?)\2\s*/g;
 let inputTypeReg = /\btype\s*=\s*(['"])([\s\S]+?)\1/;
 let tmplAttrsReg = /\$\{attrs\.([a-zA-Z_]+)\}/g;
 let tmplContentReg = /\$\{content\}/g;
-let tmplCommandAnchorReg = /\u0007\d+\u0007/g;
+let tmplCommandAnchorReg = /\u0007\d+\u0007/;
 let fileCache = Object.create(null);
 let nativeTags = (() => {
     let tags = 'a,abbr,address,area,article,aside,audio,b,base,bdi,bdo,blockquote,body,br,button,canvas,caption,cite,code,col,colgroup,data,datalist,dd,del,details,dfn,dialog,div,dl,dt,em,embed,fieldset,figcaption,figure,footer,form,h1,h2,h3,h4,h5,h6,head,header,hgroup,hr,html,i,iframe,img,input,ins,kbd,keygen,label,legend,li,link,main,map,mark,menu,menuitem,meta,meter,nav,noscript,object,ol,optgroup,option,output,p,param,pre,progress,q,rb,rp,rt,rtc,ruby,s,samp,script,section,select,small,source,span,strong,style,sub,summary,sup,table,tbody,td,template,textarea,tfoot,th,thead,time,title,tr,track,u,ul,var,video,wbr'.split(',');
@@ -44,7 +45,7 @@ let attachMap = result => {
 let toNative = (result, cmdStore, e) => {
     let attrs = result.attrs;
     let type = '';
-    let tag = result.tag;
+    let tag = result.subTag;
     if (tag == 'input') {
         let m = attrs.match(inputTypeReg);
         if (m) {
@@ -70,7 +71,7 @@ let toNative = (result, cmdStore, e) => {
                 if (ocm) {
                     return '<%if(' + ocm[2] + '){%> ' + key + ' <%}%>';
                 } else {
-                    slog.ever(chalk.red('check attribute ' + key + '=' + q + oc + q), 'at', chalk.magenta(e.shortHTMLFile));
+                    slog.ever(chalk.red('check attribute ' + key + '=' + q + oc + q), 'at', chalk.magenta(e.shortHTMLFile), 'the attribute value only support expression like ' + key + '="<%= data.' + key + ' %>" or ' + key + '="<%! data.' + key + ' %>" or ' + key + '="<%@ data.' + key + ' %>"');
                 }
             } else {
                 if (content === 'false' ||
@@ -105,7 +106,7 @@ let innerView = (result, info) => {
         return '';
     });
     let type = '';
-    let mxTagRoot = configs.mxTagViewsRoot;
+    let mxTagRoot = configs.mxGalleriesRoot;
     if (tag == 'input') {
         let m = attrs.match(inputTypeReg);
         if (m) {
@@ -134,10 +135,37 @@ let innerView = (result, info) => {
     }
     return `<${tag} ${attrs}>${result.content}</${tag}>`;
 };
+let namespaceView = (result, path) => {
+    let tag = 'div';
+    let hasTag = false;
+    let attrs = result.attrs.replace(tagReg, (m, t) => {
+        tag = t;
+        hasTag = true;
+        return '';
+    });
+    let type = '';
+    if (tag == 'input') {
+        let m = attrs.match(inputTypeReg);
+        if (m) {
+            type = m[2];
+        }
+    }
+    let allAttrs = attrMap.getAll(tag, type);
+    attrs = attrs.replace(attrNameValueReg, (m, name, q, value) => {
+        if (!allAttrs.hasOwnProperty(name) &&
+            name.indexOf('view-') !== 0 &&
+            name.indexOf('mx-') !== 0) {
+            name = 'view-' + name;
+        }
+        return ' ' + name + '=' + q + value + q;
+    });
+    attrs += ' mx-view="' + path + '"';
+    return `<${tag} ${attrs}>${result.content}</${tag}>`;
+};
 let innerInclude = (result, info) => {
     let file = '';
     if (result.subTag) {
-        let p = path.join(configs.moduleIdRemovedPath, configs.mxIncludesRoot, result.subTag);
+        let p = path.join(configs.moduleIdRemovedPath, configs.mxIncludesRoot, result.subTags.join(sep));
         for (let ext of configs.tmplFileExtNames) {
             let t = p + '.' + ext;
             if (fs.existsSync(t)) {
@@ -178,16 +206,13 @@ let innerInclude = (result, info) => {
     }
     return '';
 };
-let innerNative = result => {
-    return `<${result.subTag} ${result.attrs}></${result.subTag}>`;
-};
 module.exports = {
     process(tmpl, extInfo) {
         let cmdCache = Object.create(null);
         tmpl = tmplCmd.store(tmpl, cmdCache);
         let restore = tmpl => tmplCmd.recover(tmpl, cmdCache);
         let tokens = tmplParser(tmpl);
-        let mxTagViewsMap = configs.mxTagViewsMap;
+        let mxGalleriesMap = configs.mxGalleriesMap;
         let updateOffset = (pos, offset) => {
             let l = nodes => {
                 if (nodes) {
@@ -220,108 +245,121 @@ module.exports = {
             };
             l(tokens);
         };
+        let getTagInfo = (n, isMxTag) => {
+            let content = '',
+                attrs = '';
+            if (n.hasAttrs) {
+                attrs = tmpl.slice(n.attrsStart, n.attrsEnd);
+            }
+            if (n.hasContent) {
+                content = tmpl.slice(n.contentStart, n.contentEnd);
+            }
+            let tag = n.tag;
+            let namespace = '';
+            if (isMxTag) {
+                tag = tag.slice(3);
+            } else {
+                let temp = tag.indexOf('.');
+                namespace = tag.slice(0, temp);
+                tag = tag.slice(temp + 1);
+            }
+            let splitter = tag.indexOf('.') >= 0 ? '.' : '-';
+            let tags = tag.split(splitter);
+            if (splitter == '-' && tags.length > 1) {
+                slog.ever(chalk.red('deprecated tag:' + n.tag), 'at', chalk.magenta(extInfo.shortHTMLFile), 'use', chalk.magenta('mx-' + tags.join('.')), 'instead');
+            }
+            let result = {
+                unary: !n.hasContent,
+                namespace,
+                name: tag,
+                tag,
+                mainTag: tags.shift(),
+                splitter,
+                subTags: tags,
+                subTag: tags.join(splitter),
+                attrs,
+                content
+            };
+            return result;
+        };
+        let processMxTag = n => {
+            let result = getTagInfo(n, true);
+            let content = result.content;
+            //不能增加native tags的检测，因为组件有可能和native tag重名
+            if (!uncheckTags.hasOwnProperty(result.mainTag) &&
+                !mxGalleriesMap.hasOwnProperty(result.tag)) {
+                let cpath = path.join(configs.moduleIdRemovedPath, configs.mxGalleriesRoot);
+                cpath = path.join(cpath, 'mx-' + result.mainTag);
+                if (fs.existsSync(cpath)) {
+                    mxGalleriesMap[result.tag] = 'mx-' + result.mainTag + '/' + (result.subTag || 'index');
+                } else {
+                    uncheckTags[result.mainTag] = 1;
+                }
+            }
+            let update = false;
+            if (result.mainTag == 'view') {
+                content = innerView(result);
+                update = true;
+            } else if (result.mainTag == 'include') {
+                content = innerInclude(result, extInfo);
+                update = true;
+            } else if (result.mainTag == 'native') {
+                content = toNative(result, cmdCache, extInfo);
+                update = true;
+            } else if (mxGalleriesMap.hasOwnProperty(result.tag)) {
+                content = innerView(result, mxGalleriesMap[result.tag]);
+                update = true;
+            } else if (nativeTags.hasOwnProperty(result.tag)) {
+                result.subTag = result.tag;
+                content = toNative(result, cmdCache, extInfo);
+                update = true;
+            } else {
+                attachMap(result);
+                let tagContent = configs.mxTagProcessor(result, extInfo);
+                if (tagContent != content) {
+                    content = tagContent;
+                    update = true;
+                }
+            }
+            if (update) {
+                content = content || '';
+                tmpl = tmpl.slice(0, n.start) + content + tmpl.slice(n.end);
+                updateOffset(n.start, content.length - (n.end - n.start));
+            }
+        };
+        let processNamespaceTag = n => {
+            let result = getTagInfo(n);
+            let content = result.content;
+            let dir = path.dirname(extInfo.file).replace(configs.moduleIdRemovedPath, '') + sep;
+            let ns = result.namespace;
+            let i = dir.indexOf(sep + ns + sep);
+            if (i >= 0) {
+                dir = dir.slice(1, i) + '/' + ns;
+            } else {
+                dir = extInfo.pkgName + '/' + ns;
+            }
+            dir = dir + '/' + result.mainTag;
+            if (result.subTags.length) {
+                dir = dir + '/' + result.subTags.join('/');
+            }
+            content = namespaceView(result, dir);
+            content = content || '';
+            tmpl = tmpl.slice(0, n.start) + content + tmpl.slice(n.end);
+            updateOffset(n.start, content.length - (n.end - n.start));
+        };
         let walk = nodes => {
             if (nodes) {
                 for (let n of nodes) {
                     walk(n.children);
                     if (n.tag.indexOf('mx-') === 0) {
-                        let content = '',
-                            attrs = '';
-                        if (n.hasAttrs) {
-                            attrs = tmpl.slice(n.attrsStart, n.attrsEnd);
-                        }
-                        if (n.hasContent) {
-                            content = tmpl.slice(n.contentStart, n.contentEnd);
-                        }
-                        let tag = n.tag.slice(3);
-                        let tags = tag.split('-');
-                        let result = {
-                            unary: !n.hasContent,
-                            name: tag,
-                            tag: tag,
-                            mainTag: tags.shift(),
-                            subTag: tags.join('-'),
-                            attrs,
-                            content
-                        };
-                        //不能增加native tags的检测，因为组件有可能和native tag重名
-                        if (!uncheckTags.hasOwnProperty(result.mainTag) &&
-                            !mxTagViewsMap.hasOwnProperty(result.tag)) {
-                            let cpath = path.join(configs.moduleIdRemovedPath, configs.mxTagViewsRoot);
-                            let parts = result.tag.split('-');
-                            if (parts.length == 1) {
-                                parts.push('index');
-                            }
-                            cpath = path.join(cpath, 'mx-' + parts[0]);
-                            if (fs.existsSync(cpath)) {
-                                mxTagViewsMap[result.tag] = 'mx-' + parts.shift() + '/' + parts.join('-');
-                            } else {
-                                uncheckTags[result.mainTag] = 1;
-                            }
-                        }
-                        let update = false;
-                        if (result.mainTag == 'view') {
-                            content = innerView(result);
-                            update = true;
-                        } else if (result.mainTag == 'include') {
-                            content = innerInclude(result, extInfo);
-                            update = true;
-                        } else if (result.mainTag == 'native') {
-                            content = innerNative(result);
-                            update = true;
-                        } else if (mxTagViewsMap.hasOwnProperty(result.tag)) {
-                            content = innerView(result, mxTagViewsMap[result.tag]);
-                            update = true;
-                        } else if (!nativeTags.hasOwnProperty(result.tag)) {
-                            attachMap(result);
-                            let tagContent = configs.mxTagProcessor(result, extInfo);
-                            if (tagContent != content) {
-                                content = tagContent;
-                                update = true;
-                            }
-                        }
-                        if (update) {
-                            content = content || '';
-                            tmpl = tmpl.slice(0, n.start) + content + tmpl.slice(n.end);
-                            updateOffset(n.start, content.length - (n.end - n.start));
-                        }
+                        processMxTag(n);
+                    } else if (n.tag.indexOf('.') >= 0) {
+                        processNamespaceTag(n);
                     }
                 }
             }
         };
         walk(tokens);
-        tmpl = tmplCmd.store(tmpl, cmdCache);
-        let walkToNative = nodes => {
-            if (nodes) {
-                for (let n of nodes) {
-                    walkToNative(n.children);
-                    if (n.tag.indexOf('mx-') === 0) {
-                        let content = '',
-                            attrs = '';
-                        if (n.hasAttrs) {
-                            attrs = tmpl.slice(n.attrsStart, n.attrsEnd);
-                        }
-                        if (n.hasContent) {
-                            content = tmpl.slice(n.contentStart, n.contentEnd);
-                        }
-                        let tag = n.tag.slice(3);
-                        let result = {
-                            unary: !n.hasContent,
-                            name: tag,
-                            tag: tag,
-                            attrs,
-                            content
-                        };
-                        content = toNative(result, cmdCache, extInfo);
-                        content = content || '';
-                        tmpl = tmpl.slice(0, n.start) + content + tmpl.slice(n.end);
-                        updateOffset(n.start, content.length - (n.end - n.start));
-                    }
-                }
-            }
-        };
-        tokens = tmplParser(tmpl);
-        walkToNative(tokens);
         tmpl = restore(tmpl);
         return tmpl;
     }